more updates

+++ router/deleteItems.py
@@ -1,4 +1,5 @@
 from apis.router_api import RouterApi
+from apis.biit_api import BiitApi
 from helpers.router_helper import RouterArg
 from helpers.file_helper import FileArg
 import argparse
@@ -12,6 +13,7 @@ parser.add_argument('--router', '-r', dest='router', type=RouterArg(),
   help='(Required) One of [{}], or a custom url'.format( ", ".join( RouterArg.URLS.keys()) ), required=True)
 parser.add_argument('--input-file', '-i', dest='input', type=FileArg(), required=True, help='Input file with items to delete')
 parser.add_argument('--list', '-l', nargs='+', dest='list', required=True, help='Fields to delete. Can be multiple of {}'.format(deletableFields) )
+parser.add_argument('--verify', dest='verify', default=False, action='store_true', help='Revalidate (only applies for environment biit addresses) ')
 args = parser.parse_args()
 print("Running with args:")
 print(args)
@@ -30,7 +32,10 @@ with open(args.input) as json_file:
       api.delete_envVersion(item.get('envVerId'))
   
     if 'environment' in args.list and 'envId' in item:
-      api.delete_environment(item.get('envId'))
+      if not args.verify or not item.get('biitAddress') or not BiitApi.healthy(item.get('biitAddress')):
+        api.delete_environment(item.get('envId'))
+      else:
+        print('Skipping environment {} because biit instance was deemed healthy'.format(item.get('envId')) )
 
     if 'customer' in args.list and 'customerId' in item:
       api.delete_environment(item.get('customerId'))
